requirements.txt
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.15
PyYAML==5.3.1
Flask==1.1.1
pymysql==0.9.3
pytest==5.4.2
requests>=2.23.0
pandas>=1.1.5
numpy==1.19.5
bs4==0.0.1
boto3
botocore
mysql-connector-python>=2.2.9
lxml>=4.5.2
fsspec
s3fs
aiobotocore==1.3.0
scikit-learn>=0.23.1
statsmodels>=0.12.2
tabulate>=0.8.7
matplotlib>=3.2.1
aiohttp>=3.6.2
